Fixed Offset And Limit
Refactored DB 332 Shims and DB 341 Shims
